MetaTags and overscroll color
Phone version of the page would show a white background if you scrolled "past the end of the page", added a fallback background color to make this less harsh and noticeable.

Also added the theme-color meta tag for browsers that use this, for example safari will theme the user interface (URL bar, navigation buttons) based on this tag.